+++ task6/app/Http/Controllers/UsersController.php
@@ -26,10 +26,20 @@ class UsersController extends Controller
      */
     protected $validator;
 
+    /**
+     * @var
+     */
+    private $user;
+
+    /**
+     * @param UserRepository $repository
+     * @param UserValidator $validator
+     */
     public function __construct(UserRepository $repository, UserValidator $validator)
     {
+        $this->user = \Auth::user();
         $this->repository = $repository;
-        $this->validator  = $validator;
+        $this->validator = $validator;
     }
 
 
@@ -40,10 +50,10 @@ class UsersController extends Controller
      */
     public function index()
     {
+        if (!$this->user->is('admin')) return \Redirect::route('user.edit', array('id' => $this->user->id));
         $this->repository->pushCriteria(app('Prettus\Repository\Criteria\RequestCriteria'));
         $users = $this->repository->paginate(10);
         if (request()->wantsJson()) {
-
             return response()->json([
                 'data' => $users,
             ]);
@@ -54,8 +64,10 @@ class UsersController extends Controller
 
     public function create()
     {
+        if (!$this->user->is('admin')) return;
         return view('users.create');
     }
+
     /**
      * Store a newly created resource in storage.
      *
@@ -65,21 +77,22 @@ class UsersController extends Controller
      */
     public function store(UserCreateRequest $request)
     {
+        if (!$this->user->is('admin')) return;
         try {
             $this->validator->with($request->all())->passesOrFail(ValidatorInterface::RULE_CREATE);
             $user = $this->repository->create($request->all());
             $response = [
                 'message' => 'User created.',
-                'data'    => $user->toArray(),
+                'data' => $user->toArray(),
             ];
             if ($request->wantsJson()) {
                 return response()->json($response);
             }
-            return redirect()->back()->with('message', $response['message']);
+            return redirect()->to('user')->with('message', $response['message']);
         } catch (ValidatorException $e) {
             if ($request->wantsJson()) {
                 return response()->json([
-                    'error'   => true,
+                    'error' => true,
                     'message' => $e->getMessageBag()
                 ]);
             }
@@ -97,8 +110,8 @@ class UsersController extends Controller
      */
     public function show($id)
     {
+        if (!$this->user->is('admin')) return;
         $user = $this->repository->find($id);
-
         if (request()->wantsJson()) {
 
             return response()->json([
@@ -119,9 +132,9 @@ class UsersController extends Controller
      */
     public function edit($id)
     {
-
+        if (!$this->user->is('admin'))
+            $id = $this->user->id;
         $user = $this->repository->find($id);
-
         return view('users.edit', compact('user'));
     }
 
@@ -130,18 +143,19 @@ class UsersController extends Controller
      * Update the specified resource in storage.
      *
      * @param  UserUpdateRequest $request
-     * @param  string            $id
+     * @param  string $id
      *
      * @return Response
      */
     public function update(UserUpdateRequest $request, $id)
     {
         try {
+            if (!$this->user->is('admin')) $id = $this->user->id;
             $this->validator->with($request->all())->passesOrFail(ValidatorInterface::RULE_UPDATE);
             $user = $this->repository->update($request->all(), $id);
             $response = [
                 'message' => 'User updated.',
-                'data'    => $user->toArray(),
+                'data' => $user->toArray(),
             ];
             if ($request->wantsJson()) {
                 return response()->json($response);
@@ -150,7 +164,7 @@ class UsersController extends Controller
         } catch (ValidatorException $e) {
             if ($request->wantsJson()) {
                 return response()->json([
-                    'error'   => true,
+                    'error' => true,
                     'message' => $e->getMessageBag()
                 ]);
             }
@@ -168,6 +182,7 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
+        if (!$this->user->is('admin')) return;
         $deleted = $this->repository->delete($id);
         if (request()->wantsJson()) {
             return response()->json([
